+++ test/points_test.py
@@ -314,21 +314,24 @@ def test_points_accessors():
         p.Y = i * 3
         p.Z = i - 80
         p.Classification = i * 255 // num_points
+        p.PointSourceID = i * 255 // num_points
         points.AddPoint(p)
 
-    for i, (x, y, z, classification) in enumerate(
-        zip(points.X, points.Y, points.Z, points.Classification)
+    for i, (x, y, z, classification, point_source_id) in enumerate(
+        zip(points.X, points.Y, points.Z, points.Classification, points.PointSourceID)
     ):
         assert x == i
         assert y == i * 3
         assert z == i - 80
         assert classification == i * 255 // num_points
+        assert point_source_id == i * 255 // num_points
 
     # test slice
     assert points.X[5:10] == [x for x in range(5, 10)]
     assert points.Y[-10:] == [x * 3 for x in range(1990, 2000)]
     assert points.Z[:10] == [x - 80 for x in range(0, 10)]
     assert points.Classification[:10] == [x * 255 // num_points for x in range(0, 10)]
+    assert points.PointSourceID[:10] == [x * 255 // num_points for x in range(0, 10)]
 
     # test index setter
     for i in range(len(points)):
@@ -337,11 +340,13 @@ def test_points_accessors():
         p.Y = 30
         p.Z = 40
         p.Classification = 50
+        p.PointSourceID = 50
 
     assert all([x == 20 for x in points.X])
     assert all([y == 30 for y in points.Y])
     assert all([z == 40 for z in points.Z])
     assert all([classification == 50 for classification in points.Classification])
+    assert all([point_source_id == 50 for point_source_id in points.PointSourceID])
 
     # test iterator setter
     for p in points: